Sprint6 fix23 jd (#61)
* commit corregido

* confirmacion

* cambios stored procedure

* INGRESOS DE LOS SCRIPTS

* CONFIRMACION DE CAMBIOS ANTES DE LA FUSION

* modificacion de respuestas

* CAMBIOS STORED PROCEDURES

* MODIFICACION DE POLIZA

* MODIFICAION DE VEHICLE

* ARREGLO DE IMPRESION Y EXPORTACION

* MODIFICACION DE import de vehiculos

* cambios en autenticacion

* modificacion de brand

* agg stored procedures , cambios en msj

* cambios

* modificacion de supplierid

* commit

* modificacion de msj de seguridad y cambios en sp

* Encabezado de seguridad

* MODIFICACION DE MSJ SP

* MODIFICACION DE SP PARA VALIDACIONES DE EXCEL

* CAMBIOS DE IMPORTACION

* MODIFICACION DE IMPORTACION PARA OBTENER PLANTA

* modificacion import y exportacion

* modificacion de importacion de zonas

* MODIFICACION EN MSJ

* MODIFICACION DE EXPORTACION

* modificacion de import

* ARREGLO DE LAS IMPORTACIONES

* MODIFICACION DE CAMPOS F O M

* MODIFACION DE SP VEHICULOS Y CONTACTOS

* modificacion sp

* lote de ticket de servicios y polizas

* cambios y creacion de vehiculos facturados

* modificacion de maintenance para recibir idsupplier

* cambio de numero de orden

* modificacion de maintenance

* cambios de numero de orden

* cambios en los campos requeridos de servicios

* cambio el campor reuquerido mantenimiento

* metodos para mostrar possiblefault y assistancetype

* CAMBIOS DE SP

* CORRECION DE EXPORTACION EXCEL

* ACOMODAR DATOS DE ASISTENCIA TECNICA

* cambios de reporte de falla

* CAMBIO DE PARA SERVICE ASSISTANCE

* modificacion de exportacion de report fail

* cambios en servicios

* cambios en exportacion srg

* revision de importacion de parts

* modficacion

* encabezados de seeguridad

* mas seguridad

* IDUSER EN ZONAS

* cambios attachment

* MODIFICACION DEL PROCESAR

* modificacion de controler de vehiculos full

* commit de control de cambios

* ARREGLO DE ATTACHMENT

* SUPPLIERID PARA POLICYBY

* endpoint de usuarios asignados

* creacion de lista de reporte de fallas validados

* configuracion de licencia

* Eliminar licencias

* vEHICLEID

* arreglos en delete attachment

* DEVOLVER FECHA DE CREACION DE ARCHIVOS

* modelo y usp de elementos externos

* creacion de items

* cambio de respuesta en attachment

* Modificaciones variadas

* modificacion en importacion de tipos de polizas

* cambios en captura de numeros en importaciones

* MODIFICACION DE MOVEMENTE DETAILS

* recuperar attachment

* ajustes de bd

* modificacion de parts

* modificacion de los importar

* Captura de los Errores en importar

* modificacion de sp se seguridad

* cambios de vehiculos

* MODIFICACION DE SEGURIDAD

* MODIFICACION DE DE ACCIONES

* prueba de de exportar

* modificacion de exportar de mano de obra

* CREACION DE VALIDACION DE AÑOS VEHICULOS

* rango de año corregido

* modificacion de items

* GETONE PARA ITEMS

+++ Models/ServiceDetails.cs
@@ -18,7 +18,7 @@ namespace Models
         [Required] public Int32? ItemId { get; set; }
         [SwaggerIgnore] public String? ItemName { get; set; }
         [Required] [StringLength(1, MinimumLength = 1)] public String? Type { get; set; }
-        [Required] public  Int32? Quantity { get; set; }
+        [Required] public  decimal? Quantity { get; set; }
         [Required] public Decimal? UnitPrice { get; set; }
         [Required] public Boolean? IsExternal { get; set; }
         [Required] public Boolean? IsTax { get; set;}
